Fixes on input handling - Fixed my ft_atoll.c function - Corrected the save_input part to handle leading zeros, the mix of characters with numbers & the int limits - Improved the calculation of the length of the string represented as a number - Made everything norminette valid again

+++ libft/ft_atoll.c
@@ -6,7 +6,7 @@
 /*   By: dominique <dominique@student.codam.nl>       +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/04/29 15:34:33 by dominique     #+#    #+#                 */
-/*   Updated: 2020/04/29 15:35:25 by dominique     ########   odam.nl         */
+/*   Updated: 2020/06/04 17:27:44 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@ static long long	make_num(unsigned char *s, int sign)
 {
 	long long		num;
 	int				len;
-	int				units;
+	long long		units;
 
 	len = 0;
 	num = 0;
